AI refactor
