+++ src/app/components/transaction/votingDetails.tsx
@@ -1,9 +1,9 @@
 import { Table, TableBody, TableCell, TableContainer, TableRow, Paper, Link, Checkbox, FormControlLabel } from "@mui/material";
-import { openInNewTab } from "../utils/txUtils";
+import { openInNewTab } from "../../utils/txUtils";
 import { useEffect, useState } from "react";
-import InfoWithTooltip from "./molecules/infoHover";
-import { TOOLTIP_MESSAGES } from "../constants/infoMessages";
-import { VoteTransactionDetails } from "./types/types";
+import InfoWithTooltip from "../common/infoHover";
+import { TOOLTIP_MESSAGES } from "../../constants/infoMessages";
+import { VoteTransactionDetails } from "../types/types";
 
 
 export interface VotingDetailsProps extends VoteTransactionDetails {
