+++ src/app/components/wallet/walletDetails.tsx
