--- src/types/jest.d.ts
@@ -1,10 +1,0 @@
-import '@testing-library/jest-dom';
-
-declare global {
-  namespace jest {
-    interface Matchers<R> {
-      toBeInTheDocument(): R;
-      toHaveTextContent(text: string): R;
-    }
-  }
-}
